scripts/methods/cellpose_sam/requirements.txt
# Cellpose-SAM requirements for NucWorm benchmark
# Note: Cellpose-SAM is a recent addition to Cellpose
# May need to install from development branch or specific version

# Core Cellpose with SAM support
cellpose>=4.0.0  # Latest version with Cellpose-SAM support
torch>=1.9.0
torchvision>=0.10.0

# Additional dependencies for Cellpose-SAM
transformers>=4.20.0  # For SAM transformer backbone
timm>=0.6.0  # For vision transformer components

# Standard dependencies
numpy>=1.21.0
scipy>=1.7.0
tifffile>=2021.7.2
opencv-python-headless>=4.9.0.80
matplotlib>=3.4.0
tqdm>=4.62.0
pandas>=1.3.0
pyyaml>=5.4.0
scikit-image>=0.19.0  # For image processing

# Optional: For better performance
# accelerate>=0.20.0  # For faster training/inference
# flash-attn>=2.0.0   # For faster attention (if available)
